+++ src/main/java/com/dev/careers/config/RedisSessionConfig.java
@@ -1,0 +1,30 @@
+package com.dev.careers.config;
+
+import org.springframework.beans.factory.annotation.Value;
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.repository.configuration.EnableRedisRepositories;
+
+/**
+ * Redis 설정 및 빈 등록 (application.properties 설정값 기준) Host, Port
+ */
+@Configuration
+@EnableRedisRepositories
+public class RedisSessionConfig {
+
+    @Value("${spring.redis.host}")
+    private String host;
+
+    @Value("${spring.redis.port}")
+    private int port;
+
+    /**
+     * Netty기반 Redis Connection Factory 빈 등록 비동기 기반으로 동작함으로 Jedis보다 성능이 좋다.
+     */
+    @Bean(name = "redisConnectionFactory")
+    public RedisConnectionFactory redisConnectionFactory() {
+        return new LettuceConnectionFactory(host, port);
+    }
+}
