+++ src/main/resources/application-dev.properties
@@ -4,3 +4,7 @@ spring.datasource.password=1234
 
 spring.redis.host=localhost
 spring.redis.port=6379
+
+spring.redis.cache.host=localhost
+spring.redis.cache.port=6380
+spring.redis.cache.ttl=1
