+++ src/main/resources/application-prod.properties
@@ -4,3 +4,7 @@ spring.datasource.password=${MYSQL_DB_PASSWORD}
 
 spring.redis.host=${REDIS_HOST}
 spring.redis.port=${REDIS_PORT}
+
+string.redis.cache.host=${REDIS_CACHE_HOST}
+string.redis.cache.port=${REDIS_CACHE_PORT}
+spring.redis.cache.ttl=${REDIS_CACHE_TTL}
